EES-4192 - misc additional changes made whilst performance testing EES-4192.
Removed "ProcessUploads" overload in Processor.cs due to startup warnings in Functions project.

Added config to allow importer to dequeue messages faster.

Amended adminTableBuilderQuery.test.ts and publicTableBuilderQuery.test.ts to be more closely aligned in their behaviour after one was left working and the other was no longer working. Hopefully this will be easier to maintain.
