+++ src/GovUk.Education.ExploreEducationStatistics.Data.Processor.Tests/Functions/ProcessorStage2Tests.cs
@@ -266,8 +266,7 @@ public class ProcessorStage2Tests
         await function.ProcessUploads(
             importMessage,
             new ExecutionContext(),
-            importStagesMessageQueue.Object,
-            rethrowExceptions: true);
+            importStagesMessageQueue.Object);
 
         VerifyAllMocks(blobStorageService, importStagesMessageQueue);
 
@@ -438,6 +437,7 @@ public class ProcessorStage2Tests
         return new Processor.Functions.Processor(
             dataImportService ?? Mock.Of<IDataImportService>(Strict),
             processorService ?? Mock.Of<IProcessorService>(Strict),
-            Mock.Of<ILogger<Processor.Functions.Processor>>());
+            Mock.Of<ILogger<Processor.Functions.Processor>>(),
+            rethrowExceptions: true);
     }
 }
