+++ src/GovUk.Education.ExploreEducationStatistics.Data.Processor/host.json
@@ -14,7 +14,8 @@
     "queues": {
       "batchSize": 6,
       "maxDequeueCount": 3,
-      "newBatchThreshold": 1
+      "newBatchThreshold": 1,
+      "maxPollingInterval": "00:00:02"
     }
   }
 }
